refactor: update wins
